im gonna be focusing on teaching you self hypnosis now

+++ backend/school/response.py
@@ -80,8 +80,8 @@ def create_school(request: WSGIRequest):
         communicator.delete()
         return JsonResponse({
             "status": "Error",
-            "error": "This user already exists",
-        }, status=400)
+            "error": "Failed to pair user data object to user",
+        }, status=500)
 
     try:
         school = School.objects.create(
@@ -108,19 +108,55 @@ def create_school(request: WSGIRequest):
 
 @login_required
 @wrappers.require_role(["organizer", "school"])
-@require_http_methods(["DELETE"])
-def delete_school(request: WSGIRequest, school_id):
+@require_http_methods(["DELETE", "PATCH"])
+def manage_school(request: WSGIRequest, school_id):
     user_data = UserData.objects.get(user=request.user)
     school = School.objects.get(id=school_id)
-    if user_data.role == "school" and school.communicator != request.user:
+    if request.method == "DELETE":
+        if user_data.role == "school" and school.communicator != request.user:
+            return JsonResponse({
+                "status": "Error",
+                "error": "You do not have permission to delete this school",
+            }, status=403)
+
+        school.communicator.delete()
+
+        return JsonResponse({
+            "status": "Ok",
+            "error": None,
+        }, status=200)
+
+    try:
+        body = json.loads(request.body.decode('utf-8'))
+    except JSONDecodeError:
         return JsonResponse({
             "status": "Error",
-            "error": "You do not have permission to delete this school",
-        }, status=403)
+            "error": "Invalid request body",
+        }, status=400)
 
-    school.communicator.delete()
+    school_obj = School.objects.get(id=school_id)
+    if "school_name" in body.keys():
+        school_obj.name = body["school_name"]
+    if "address" in body.keys():
+        school.address = body["address"]
+    school.save()
+
+    if "username" in body.keys():
+        school.communicator.username = body["username"]
+    if "password" in body.keys():
+        school.communicator.set_password(body["password"])
+    if "email" in body.keys():
+        school.communicator.email = body["email"]
+    school.communicator.save()
+
+    user_data = UserData.objects.get(user=school.communicator)
+    if "display_name" in body.keys():
+        user_data.display_name = body["display_name"]
+    user_data.save()
+
 
     return JsonResponse({
         "status": "Ok",
         "error": None,
+        "modified": modules.django_model_operations.model_to_dict(school),
     }, status=200)
